Reset global state of jail_code to test django_integration_utils
